requirements.txt
Pillow==10.3.0
PyMuPDF==1.24.1
anthropic==0.25.1
numpy==1.26.4
opencv-python==4.9.0.80
pyodbc
azure-identity
pytest
setuptools
pytesseract
reportlab